backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.3.0
openpyxl==3.1.2
WeasyPrint==62.3
django-filter==24.2
